Introduced PRIVATE_INSTANCE const and associated logic to clarify PRIVATE_INSTANCE changes.  Use redirection to bypass terms page. Now all original tests pass

+++ app/controllers/user_controller.rb
@@ -79,6 +79,10 @@ class UserController < ApplicationController
         elsif @user.new_record?
           # Save the user record
           session[:new_user] = @user
+
+          if PRIVATE_INSTANCE 
+            redirect_to :action => :save
+          end
         end
       else
         # Not logged in, so redirect to the login page
@@ -123,13 +127,6 @@ class UserController < ApplicationController
       end
     else
       @user = session.delete(:new_user)
-      # In order for terms bypass to work we need to create user from params, not sesh
-      # TODO:  Verify correctness in face of OAuth/OpenID logic
-      if !@user and params[:user]
-        @user = User.new(params[:user])
-        @user.status = "pending"
-      end
-	
 
       if Acl.no_account_creation(request.remote_ip, @user.email.split("@").last)
         render :action => 'blocked'
@@ -153,10 +150,12 @@ class UserController < ApplicationController
             successful_login(@user)
           else
             flash[:notice] = t 'user.new.flash create success message', :email => @user.email
-            session[:token] = @user.tokens.create.token
-
-            Notifier.signup_confirm(@user, @user.tokens.create(:referer => session.delete(:referer))).deliver
-
+            if PRIVATE_INSTANCE
+              Notifier.signup_confirm(@user, nil).deliver
+            else
+              session[:token] = @user.tokens.create.token
+              Notifier.signup_confirm(@user, @user.tokens.create(:referer => session.delete(:referer))).deliver
+            end
             redirect_to :action => 'login', :referer => params[:referer]
           end
         else
